+++ packages/integration-sdk-core/src/types/storage.ts
@@ -1,9 +1,5 @@
 import { Entity } from './entity';
-import {
-  GraphObjectFilter,
-  GraphObjectIteratee,
-  GraphObjectLookupKey,
-} from './jobState';
+import { GraphObjectFilter, GraphObjectIteratee } from './jobState';
 import { Relationship } from './relationship';
 import { GraphObjectIndexMetadata } from '../types/step';
 
@@ -30,7 +26,7 @@ export interface GraphObjectStore {
     onRelationshipsFlushed?: (relationships: Relationship[]) => Promise<void>,
   ): Promise<void>;
 
-  getEntity({ _key, _type }: GraphObjectLookupKey): Promise<Entity>;
+  findEntity(_key: string): Promise<Entity | undefined>;
 
   iterateEntities<T extends Entity = Entity>(
     filter: GraphObjectFilter,